[HMTB-2078] Removing fuel benefit value from confirmation page.
First part of the car / fuel benefit confirmation story: we just
removed the value of the benefit from the table and all throughout
the page.

NOTE: We added extra test for making sure that nowhere else the amounts
and percentages are displayed in the page. This was because of a
conversation with the BA on the logic of the confirmation page.