[HMTB-2078] Call to calculation service removed from pre-confirmation page.
Since the PAYE add car/fuel benefit does the calculation
automatically, it is not needed anymore the gross amounts
calculation through the explicit invocation of the calculation
service.

Similarly the user's keyStore would not contain anymore
amounts but only the details of the benefit to be added.

All assumptions / assertions / verifications about calculation
services and amounts on keyStore have been removed from tests
as they are no longer relevant.