Now building the header carrier with the timestamp from the request (or System.nanoTime if the request timestamp is missing)
I have changed Connector.httpPostF to accept an Option[TBody] for the body rather than accepting null. I have also re-named the type parameters to be more meaningful and re-ordered them so that the response type is the first param, similar to things like FunctionN, where the return type param comes first. Still need to fix up the specs.

Explicitly declaring the type params on the httpPostF call in the auth connector so that the manifests are correctly created by scala.

Needed to update a lot of the specs because header carriers now have a timestamp and two carriers created at different timestamps won't compare equal anymore.

Change-Id: I39ccbc1b70d15a5805d63e038ff11e121feb04f4