App configuration managed by puppet.
Replaced all the environment specific conf files with a single default
file, application.conf (this file contains the values that were in the
'local' configuration, so running the apps locally should be unaffected).

In all other environments, system property overrides are supplied on the
command line.

The environment specific properties are managed by puppet, see:

https://github.gds/HMRC/puppet/pull/387

Also replaced the environment specific startup scripts with a single
script, as the start command is now common across environments.