[HMTB-2830] I've changed WithSessionTimeoutValidation to accept an authenticationProvider as a parameter. In Actions this parameter will be satisfied with the authentication provider from the regime that the controller method specified to AuthorisedFor. If WithSessionTimeoutValidation detects that the session has timed out it will use the provided authentication provider to redirect to the appropriate login page.
Previously we were using AnyAuthenticationProvider in this case, which attempted to work out which login page was needed by looking at a string value held on the session. However, given the session had timed out, that parameter would not be available and it would use the default login page, i.e. business tax login.

Change-Id: Id4c328b361bb4714bb3f0e90055532ef5b49bd46