Update version of auth-test lib to align with current auth service
Until now the auth-test library has not been maintained by the Auth team and as a consequence we were not aware of that library getting out of sync with latest API requirements of the auth microservice. The next version of that service which we will deploy to dev environments later this week (Aug 4/5) is less forgiving in receiving requests which break the contract, so your tests would fail after that deployment.

We recommend to merge this PR in the upcoming days to avoid disruption when that deployment happens at the end of this week. Please watch the jenkins build after the merge and report any issues to the Auth team. Due to timing constraints and the need to produce 19 PRs for 7 teams, we did not actually test the change for all 19 affected services. We ran tests for a selection of them and are fairly confident that it will work for your service, too. In the unlikely case it does not, please revert the commit and report the issue to the Auth team.

In general this is a low-risk PR as it only affects your integration tests, not your production code.